Revert "correção bug setor e ajustes"
