+++ src/br/com/setor/SetorController.java
@@ -98,13 +98,7 @@ public class SetorController implements Initializable {
     }
     
     public int excluirSetor() throws SQLException{
-        Setor setorDelete;
-        if(tabelaSetor.getSelectionModel().getSelectedItem() == null){
-            functions.mensagemPadrao("Favor selecionar setor para exclusão!");
-            return Functions.FAILURE;
-        }else{
-            setorDelete = tabelaSetor.getSelectionModel().getSelectedItem();
-        }
+        Setor setorDelete = tabelaSetor.getSelectionModel().getSelectedItem();
         try {
             if (setorDelete.getQuant()>0 ){
                 functions.mensagemPadrao("Setor possui pessoas cadastradas. Não será possível realizar a exclusão!");
