+++ src/br/com/softtalk/MenuController.java
@@ -5,7 +5,6 @@
  */
 package br.com.softtalk;
 
-import br.com.pessoa.Pessoa;
 import br.com.setor.Setor;
 import java.io.IOException;
 import java.net.URL;
@@ -23,6 +22,7 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.image.ImageView;
+import javafx.scene.text.Text;
 
 /**
  * FXML Controller class
@@ -56,13 +56,8 @@ public class MenuController implements Initializable {
 
     @FXML
     void perfilAction(ActionEvent event) {
-        Parent fxmlLoader;
-        try {
-            fxmlLoader = FXMLLoader.load(Pessoa.class.getResource("Pessoa.fxml"));
-            SoftTalk.stage.setScene(new Scene(fxmlLoader));
-        } catch (IOException ex) {
-            Logger.getLogger(MenuController.class.getName()).log(Level.SEVERE, null, ex);
-        }
+        SoftTalkController softTalk = new SoftTalkController();
+        softTalk.abrirPerfil();
     }
 
     @FXML
