+++ src/br/com/softtalk/SoftTalk.fxml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.AnchorPane?>
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="500.0" prefWidth="900.0" style="-fx-background-color: #eeeeee;" stylesheets="@Anc.css" xmlns="http://javafx.com/javafx/8.0.141" xmlns:fx="http://javafx.com/fxml/1" fx:controller="br.com.softtalk.SoftTalkController">
+   <children>
+      <fx:include fx:id="testePesquisa" source="Menu.fxml" />
+   </children>
+</AnchorPane>
