--- src/br/com/telainicial/TelaInicial.fxml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import java.net.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="600.0" styleClass="mainFxmlClass" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8" fx:controller="br.com.telainicial.TelaInicialController">
-    <stylesheets>
-        <URL value="@telainicial.css" />
-    </stylesheets>
-   <children>
-      <fx:include fx:id="testePesquisa" source="../softtalk/Menu.fxml" />
-      <AnchorPane prefHeight="118.0" prefWidth="600.0" />
-   </children>
-</AnchorPane>
