--- src/br/com/telainicial/telainicial.css
@@ -1,7 +1,0 @@
-/*
- * Empty Stylesheet file.
- */
-
-.mainFxmlClass {
-
-}
